Removes unused piece of code from wp_template().
This was supposed to keep template files from being called directly.
But it doesn't work. Given the fact that the template file calls
a function that is not defined (wp_template) it will die anyway.
The only time this doesn't work is when functions.php is called.
Hence, I use the code at the top of functions.php.